Update README.md
hello there.

this is sindhu.
i am a dentist by profession. i would like to enter into the world of programming/coding..
its coz i dnt want to be bound at a place and want to move around(different countries), come out of the box which unfortunately is not possible with my profession,as my degree is not valid in most of the countriesand its inflexibility to moe around.
i found the lifestyle of my friends from other profession very flexible and interesting and i was so much attracted and fascinated with the black magic screen. then i decided this would be the way to my dreams and made up my mind to learn programming. now as am just a beginner and would love to use help/tips from any.
this is all about me.

thank you!